fixed64: hand-fix Fixed*Fixed multiplies (per-site raw-product intent)
The bare fixed*fixed multiplies (kept a compile error on purpose) are all
raw products in the engine; fixed each to its explicit intent:
  - intMagnitude * fixed  ((dist>>FRACBITS)*finecosine, (radius>>16)*
    (height>>16), (height>>3)*(velz>>16), topdist*plane.c, Distance*
    FRACUNIT): (int) cast on the integer-magnitude operand -> int*fixed
    raw product, same value as the non-strict build.
  - genuine raw fixed*fixed / fixed*fixed/fixed (am_map FTOM, A_Mushroom
    z-aim, decal spread r*ldx/wallsize): the tested Scale(a,b,c) helper
    (a*b/c in 128-bit), with c=1 for the plain products.
Behavior-preserving; flag-off green; tests green (104).

+++ src/zandronum/src/botcommands.cpp
@@ -1337,8 +1337,8 @@ static void botcmd_PathToGoal( CSkullBot *pBot )
 
 	// We don't need GoalPos anymore, so we can corrupt it! KEKE!
 	Angle = Angle >> ANGLETOFINESHIFT;
-	GoalPos.x = pBot->GetPlayer( )->mo->x + ( USERANGE >> FRACBITS ) * finecosine[Angle];
-	GoalPos.y = pBot->GetPlayer( )->mo->y + ( USERANGE >> FRACBITS ) * finesine[Angle];
+	GoalPos.x = pBot->GetPlayer( )->mo->x + (int)( USERANGE >> FRACBITS ) * finecosine[Angle];
+	GoalPos.y = pBot->GetPlayer( )->mo->y + (int)( USERANGE >> FRACBITS ) * finesine[Angle];
 
 	ulFlags = BOTPATH_TryWalk( pBot->GetPlayer( )->mo, pBot->GetPlayer( )->mo->x, pBot->GetPlayer( )->mo->y, pBot->GetPlayer( )->mo->z, GoalPos.x, GoalPos.y );
 	if ( ulFlags & BOTPATH_JUMPABLELEDGE )
@@ -1453,8 +1453,8 @@ static void botcmd_PathToLastKnownEnemyPosition( CSkullBot *pBot )
 
 	// We don't need GoalPos anymore, so we can corrupt it! KEKE!
 	Angle = Angle >> ANGLETOFINESHIFT;
-	GoalPos.x = pBot->GetPlayer( )->mo->x + ( USERANGE >> FRACBITS ) * finecosine[Angle];
-	GoalPos.y = pBot->GetPlayer( )->mo->y + ( USERANGE >> FRACBITS ) * finesine[Angle];
+	GoalPos.x = pBot->GetPlayer( )->mo->x + (int)( USERANGE >> FRACBITS ) * finecosine[Angle];
+	GoalPos.y = pBot->GetPlayer( )->mo->y + (int)( USERANGE >> FRACBITS ) * finesine[Angle];
 
 	ulFlags = BOTPATH_TryWalk( pBot->GetPlayer( )->mo, pBot->GetPlayer( )->mo->x, pBot->GetPlayer( )->mo->y, pBot->GetPlayer( )->mo->z, GoalPos.x, GoalPos.y );
 	if ( ulFlags & BOTPATH_JUMPABLELEDGE )
@@ -1585,8 +1585,8 @@ static void botcmd_Roam( CSkullBot *pBot )
 
 	// We don't need GoalPos anymore, so we can corrupt it! KEKE!
 	Angle = Angle >> ANGLETOFINESHIFT;
-	GoalPos.x = pBot->GetPlayer( )->mo->x + ( USERANGE >> FRACBITS ) * finecosine[Angle];
-	GoalPos.y = pBot->GetPlayer( )->mo->y + ( USERANGE >> FRACBITS ) * finesine[Angle];
+	GoalPos.x = pBot->GetPlayer( )->mo->x + (int)( USERANGE >> FRACBITS ) * finecosine[Angle];
+	GoalPos.y = pBot->GetPlayer( )->mo->y + (int)( USERANGE >> FRACBITS ) * finesine[Angle];
 
 	ulFlags = BOTPATH_TryWalk( pBot->GetPlayer( )->mo, pBot->GetPlayer( )->mo->x, pBot->GetPlayer( )->mo->y, pBot->GetPlayer( )->mo->z, GoalPos.x, GoalPos.y );
 	if ( ulFlags & BOTPATH_JUMPABLELEDGE )
